+++ src/Adapter/Start/Start.mjs
@@ -1,5 +1,5 @@
 /** @typedef {import("../Semester/Semester.mjs").Semester} Semester */
 
 /**
- * @typedef {{semesters: Semester[], "min-password-length": number, "hash-password-on-client": boolean, "hash-password-on-client-algorithm": string | null, "hash-password-on-client-radix": number | null}} Start
+ * @typedef {{semesters: Semester[], "min-password-length": number, "hash-password-on-client": boolean, "hash-password-on-client-algorithm": string | null, "hash-password-on-client-radix": number | null, "hash-password-on-client-padding-end": boolean | null, "hash-password-on-client-padding-length": number | null, "hash-password-on-client-padding-character": string | null}} Start
  */