+++ src/Service/Password/Command/HashPasswordCommand.mjs
@@ -1,28 +1,18 @@
-/** @typedef {import("../../../Libs/flux-hash-api/src/Adapter/Api/HashApi.mjs").HashApi} HashApi */
 /** @typedef {import("../../../Adapter/Start/Start.mjs").Start} Start */
 
 export class HashPasswordCommand {
     /**
-     * @type {HashApi}
-     */
-    #hash_api;
-
-    /**
-     * @param {HashApi} hash_api
      * @returns {HashPasswordCommand}
      */
-    static new(hash_api) {
-        return new this(
-            hash_api
-        );
+    static new() {
+        return new this();
     }
 
     /**
-     * @param {HashApi} hash_api
      * @private
      */
-    constructor(hash_api) {
-        this.#hash_api = hash_api;
+    constructor() {
+
     }
 
     /**
@@ -35,10 +25,6 @@ export class HashPasswordCommand {
             return password;
         }
 
-        return this.#hash_api.generateHash(
-            password,
-            start["hash-password-on-client-algorithm"],
-            start["hash-password-on-client-radix"]
-        );
+        return Array.from(new Uint8Array(await crypto.subtle.digest(start["hash-password-on-client-algorithm"] ?? "SHA-256", new TextEncoder().encode(password)))).map(x => x.toString(start["hash-password-on-client-radix"] ?? 16)[start["hash-password-on-client-padding-end"] ?? false ? "padEnd" : "padStart"](start["hash-password-on-client-padding-length"] ?? 2, start["hash-password-on-client-padding-character"] ?? "0")).join("");
     }
 }