+++ src/Service/Password/Port/PasswordService.mjs
@@ -1,28 +1,18 @@
-/** @typedef {import("../../../Libs/flux-hash-api/src/Adapter/Api/HashApi.mjs").HashApi} HashApi */
 /** @typedef {import("../../../Adapter/Start/Start.mjs").Start} Start */
 
 export class PasswordService {
     /**
-     * @type {HashApi}
-     */
-    #hash_api;
-
-    /**
-     * @param {HashApi} hash_api
      * @returns {PasswordService}
      */
-    static new(hash_api) {
-        return new this(
-            hash_api
-        );
+    static new() {
+        return new this();
     }
 
     /**
-     * @param {HashApi} hash_api
      * @private
      */
-    constructor(hash_api) {
-        this.#hash_api = hash_api;
+    constructor() {
+
     }
 
     /**
@@ -31,9 +21,7 @@ export class PasswordService {
      * @returns {Promise<string>}
      */
     async hashPassword(password, start) {
-        return (await import("../Command/HashPasswordCommand.mjs")).HashPasswordCommand.new(
-            this.#hash_api
-        )
+        return (await import("../Command/HashPasswordCommand.mjs")).HashPasswordCommand.new()
             .hashPassword(
                 password,
                 start